@@ -9,7 +9,8 @@
     "minimum-stability" : "beta",
     "require": {
         "php": ">=5.3.0",
-        "symfony/symfony": "~2.1"
+        "symfony/symfony": "~2.8|~3.0",
+        "doctrine/inflector": "~1.0"
     },
     "require-dev": {
         "jms/serializer-bundle": "dev-master"
