@@ -13,11 +13,14 @@
 
 namespace SimpleThings\FormSerializerBundle\Controller;
 
+use SimpleThings\FormSerializerBundle\Serializer\FormSerializer;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormTypeInterface;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBag;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
 
 /**
  * Form trait helper implementing Restful Controllers
@@ -49,7 +52,7 @@ trait RestFormHelper
     protected $form;
 
     /**
-     * @return SimpleThings\FormSerializerBundle\Serializer\FormSerializer
+     * @return FormSerializer
      */
     protected function getFormSerializer()
     {
@@ -75,17 +78,18 @@ trait RestFormHelper
     /**
      * Create and return a form (failure) response based on the HTTP Response format.
      *
-     * @param FormInterface $data
+     * @param FormInterface $form
      * @param array         $variables Additional data that is passed to an HTML view.
      *
      * @return Response
+     * @throws \InvalidArgumentException
      */
     protected function renderForm(FormInterface $form = null, array $variables = array())
     {
         $form   = $form ?: $this->form;
         $format = $this->getRequest()->getRequestFormat();
 
-        if ($format === "html") {
+        if ($format === 'html') {
             $variables['form'] = $form->createView();
             $variables['data'] = $form->getData();
 
@@ -120,13 +124,13 @@ trait RestFormHelper
      */
     protected function flash()
     {
-        if ($this->getRequest()->getRequestFormat() !== "html") {
+        if ($this->getRequest()->getRequestFormat() !== 'html') {
             return new FlashBag; // dummy flush-bag, to keep the fluent
         }
 
         $args = func_get_args();
 
-        if (count($args) == 2) {
+        if (count($args) === 2) {
             $this->get('session')->getFlashBag()->add($args[0], $args[1]);
         }
 
@@ -145,14 +149,15 @@ trait RestFormHelper
      * @param bool   $absolute
      *
      * @return RedirectResponse
+     * @throws \InvalidArgumentException
      */
     protected function redirectRoute($routeName, $parameters = array(), $statusCode = 301, $absolute = false)
     {
         $link = $this->generateUrl($routeName, $parameters, $absolute);
 
         if ($statusCode === 201 || $statusCode === 204) {
-            if ($this->getRequest()->getRequestFormat() !== "html") {
-                return new Response("", $statusCode, array("Location" => $link));
+            if ($this->getRequest()->getRequestFormat() !== 'html') {
+                return new Response('', $statusCode, array('Location' => $link));
             }
             $statusCode = 301;
         }
@@ -167,12 +172,12 @@ trait RestFormHelper
      * @param mixed             $data
      * @param array             $options
      *
-     * @return FormInterace
+     * @return FormInterface
      */
     protected function bindForm(FormTypeInterface $type, $data = null, $options = array())
     {
         $form = $this->createForm($type, $data, $options);
-        $form->bind($this->getRequest());
+        $form->submit($this->getRequest());
 
         return $form;
     }
